refactor: move calc functions into new legend labels file
